+++ pkg/cloud/vsphere/services/govmomi/service.go
@@ -457,117 +457,3 @@ func (vms *VMService) deleteTags(ctx *context.MachineContext) error {
 	}
 	return nil
 }
-
-func (vms *VMService) reconcileIPAM(ctx *context.MachineContext, vm infrav1.VirtualMachine) (bool, error) {
-
-	if ctx.VSphereMachine == nil {
-		ctx.Logger.V(4).Info("machine infrastructure missing")
-		return false, nil
-	}
-
-	// Create a copy of the devices
-	devices := make([]infrav1.NetworkDeviceSpec, len(ctx.VSphereMachine.Spec.Network.Devices))
-	for i := range ctx.VSphereMachine.Spec.Network.Devices {
-		ctx.VSphereMachine.Spec.Network.Devices[i].DeepCopyInto(&devices[i])
-	}
-
-	// Find all devices that should get a static IP
-	staticDevicesWithoutIPs := make([]*infrav1.NetworkDeviceSpec, 0)
-	for i, device := range devices {
-		if !device.DHCP4 && !device.DHCP6 && len(device.IPAddrs) == 0 {
-			staticDevicesWithoutIPs = append(staticDevicesWithoutIPs, &devices[i])
-		}
-	}
-
-	// If no static devices need IPs then nothing to do
-	if len(staticDevicesWithoutIPs) == 0 {
-		return true, nil
-	}
-
-	// TODO(thorsteinnth) Should we be doing the ip<->mac linking?
-	// If we do, then we need to add the MAC address to the device spec, making it static, so that on re-clone
-	// the machine comes up with the same mac address - preserving the ip<->mac link in the IPAM provider
-	// At that point, are we guaranteed that it will remain unique, and that we do not need some "MAC address management service"?
-	// i.e. can we mix generated MACs and static MACs in the same environment.
-
-	/*
-		// Assign MAC addresses to the static devices
-		// NOTE: These MAC addresses should have been generated for us at this point, since the VM has already been cloned.
-		// We add the MAC addresses explicitly to the spec here, so that on eventual VM deletion from vSphere and re-clone,
-		// the MAC addresses will stay the same.
-		if len(vm.Network) != len(devices) {
-			ctx.Logger.V(6).Info("reenqueue to wait for network status") // TODO Needed?
-			return false, nil
-		}
-		for i := range devices {
-			devices[i].MACAddr = vm.Network[i].MACAddr
-		}*/
-
-	//staticMacAddresses := macAddresses[0:len(staticDevicesWithoutIPs)]
-
-	// Determine zone
-	if ctx.VSphereCluster == nil {
-		ctx.Logger.V(4).Info("cluster infrastructure missing")
-		return false, nil
-	}
-	isManagementZone := ctx.VSphereCluster.Spec.CloudProviderConfiguration.Labels.Zone == util.ManagementZoneName
-
-	// Determine network types for each device
-	networkTypeDeviceMap := make(map[ipam.NetworkType][]*infrav1.NetworkDeviceSpec)
-	for i := range staticDevicesWithoutIPs {
-		networkType, err := util.GetNetworkType(*ctx.VSphereMachine, isManagementZone, staticDevicesWithoutIPs[i].NetworkName)
-		if err != nil {
-			return false, errors.Wrapf(err, "could not get network type")
-		}
-		networkTypeDeviceMap[networkType] = append(networkTypeDeviceMap[networkType], staticDevicesWithoutIPs[i])
-	}
-
-	agent, err := util.GetIPAMAgent(ctx.Logger, ctx.Client)
-	if err != nil {
-		return false, errors.Wrap(err, "could not get IPAM agent")
-	}
-
-	for networkType, networkTypeDevices := range networkTypeDeviceMap {
-		// TODO Skipping the MAC address linking for now
-		reservations, err := agent.ReserveIPs(networkType, ipam.IPv4, len(networkTypeDevices), nil)
-		if err != nil {
-			return false, errors.Wrapf(err, "could not reserve IPs for network type %q", string(networkType))
-		}
-		if err := assignReservationsToDevices(reservations, networkTypeDevices); err != nil {
-			// Let's try to release the reservations
-			ips := getReservationIPs(reservations)
-			if err := agent.ReleaseIPs(networkType, ips); err != nil {
-				ctx.Logger.Error(err, "failed to assign IP reservations, could not release IPs", "IPs", ips)
-			}
-			return false, errors.Wrap(err, "could not assign IP reservations to devices")
-		}
-		// Assign the modified devices
-		ctx.VSphereMachine.Spec.Network.Devices = devices
-		// Add annotation marking machine as IPAM managed
-		ctx.VSphereMachine.Annotations[util.IPAMManagedAnnotationKey] = "true"
-	}
-
-	return true, nil
-}
-
-func assignReservationsToDevices(reservations []ipam.IPAddressReservation, devices []*infrav1.NetworkDeviceSpec) error {
-	if len(reservations) != len(devices) {
-		return fmt.Errorf("unexpected number of reservations %d, wanted %d", len(reservations), len(devices))
-	}
-	for i := range devices {
-		reservation := reservations[i]
-		devices[i].IPAddrs = append(devices[i].IPAddrs, reservation.Address)
-		devices[i].Nameservers = reservation.NetworkConfig.NameServers
-		devices[i].Gateway4 = reservation.NetworkConfig.DefaultGateway
-		devices[i].SearchDomains = reservation.NetworkConfig.DomainSearch
-	}
-	return nil
-}
-
-func getReservationIPs(reservations []ipam.IPAddressReservation) []string {
-	ips := make([]string, 0)
-	for _, res := range reservations {
-		ips = append(ips, res.Address)
-	}
-	return ips
-}
